Fix file setting issue
It seems that the "files" table should be included in the form, as each file container has a unique GUID name assigned to it, so the changed settings cannot be applied since Tixati client doesn't know which exact file you're applying the changes to. This change fixed the issue for me and now I'm able to have the desired changes applied to the selected files.